pda: drop the redundant s16 narrowing on fn_8003F464's clip rect
The two fn_800FE38C arguments were wrapped in (s16), which emits an extsh
retail does not have. 85.53 -> 85.61.

Verified against the target asm rather than assumed: there is no extsh at
that call site. Twenty other `(s16)(*(s16*)...)` sites remain in this file
and some may be the same redundancy, but removing a narrowing cast changes
semantics when the value can exceed 16 bits, so each needs checking against
its own target asm before being touched. Left as a lead.

+++ src/game/menu/pda_range_80037158.c
@@ -6349,10 +6349,10 @@ void fn_8003F464(PdaSprite* alphaSprite, PdaSprite* sprite)
     case 0xd6f:
         pdaPlaceSprite(alphaSprite, sprite);
         fn_800FE38C(
-            (s16)(*(s16*)(lbl_802EF0A8 + 0x17b52) -
-                  *(s16*)(lbl_802EF0A8 + sprite->eventId * 0x1c + 2)),
-            (s16)(*(s16*)(lbl_802EF0A8 + 0x17b54) -
-                  *(s16*)(lbl_802EF0A8 + sprite->eventId * 0x1c + 4)),
+            *(s16*)(lbl_802EF0A8 + 0x17b52) -
+                *(s16*)(lbl_802EF0A8 + sprite->eventId * 0x1c + 2),
+            *(s16*)(lbl_802EF0A8 + 0x17b54) -
+                *(s16*)(lbl_802EF0A8 + sprite->eventId * 0x1c + 4),
             *(s16*)(lbl_802EF0A8 + 0x17b56),
             *(s16*)(lbl_802EF0A8 + 0x17b58));
         if (*(s8*)(S + 0x15d) != 4) {
